shashank_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.4.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohappyeyeballs=2.5.0=pypi_0
aiohttp=3.11.13=pypi_0
aiosignal=1.3.2=pypi_0
airportsdata=20250224=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
astor=0.8.1=pypi_0
asttokens=3.0.0=pypi_0
attrs=25.1.0=pypi_0
auto-gptq=0.7.1=pypi_0
av=14.2.0=pypi_0
beartype=0.14.1=pypi_0
better-abc=0.0.3=pypi_0
bitsandbytes=0.45.3=pypi_0
blake3=1.0.4=pypi_0
boto3=1.37.10=pypi_0
botocore=1.37.10=pypi_0
braceexpand=0.1.7=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
cachetools=5.5.2=pypi_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cloudpickle=3.1.1=pypi_0
compressed-tensors=0.9.2=pypi_0
contourpy=1.3.1=pypi_0
cpm-kernels=1.0.11=pypi_0
cupy-cuda12x=13.4.1=pypi_0
cycler=0.12.1=pypi_0
datasets=3.3.2=pypi_0
decorator=5.2.1=pypi_0
deepspeed=0.16.4=pypi_0
depyf=0.18.0=pypi_0
diffusers=0.32.2=pypi_0
dill=0.3.8=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
dnspython=2.7.0=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.8.1=pypi_0
email-validator=2.2.0=pypi_0
executing=2.2.0=pypi_0
fancy-einsum=0.0.3=pypi_0
fastapi=0.115.11=pypi_0
fastapi-cli=0.0.7=pypi_0
fastrlock=0.8.3=pypi_0
ffmpy=0.5.0=pypi_0
filelock=3.17.0=pypi_0
fonttools=4.56.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
gekko=1.2.1=pypi_0
gguf=0.10.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gradio=5.20.1=pypi_0
gradio-client=1.7.2=pypi_0
groovy=0.1.2=pypi_0
h11=0.14.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.29.2=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
imageio-ffmpeg=0.6.0=pypi_0
importlib-metadata=8.6.1=pypi_0
interegular=0.3.3=pypi_0
ipdb=0.13.13=pypi_0
ipython=9.0.2=pypi_0
ipython-pygments-lexers=1.1.1=pypi_0
jaxtyping=0.3.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.9.0=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
kiwisolver=1.4.8=pypi_0
lark=1.2.2=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llvmlite=0.43.0=pypi_0
lm-format-enforcer=0.10.11=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.1=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mistral-common=1.5.4=pypi_0
moviepy=2.1.2=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
msgspec=0.19.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=pypi_0
ninja=1.11.1.3=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
openai=1.65.5=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openssl=3.0.16=h5eee18b_0
optimum=1.24.0=pypi_0
orjson=3.10.15=pypi_0
outlines=0.1.11=pypi_0
outlines-core=0.1.26=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
partial-json-parser=0.2.1.1.post5=pypi_0
peft=0.15.0=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.4.0=pypi_0
pip=25.0=py311h06a4308_0
platformdirs=4.3.7=pypi_0
proglog=0.1.10=pypi_0
prometheus-client=0.21.1=pypi_0
prometheus-fastapi-instrumentator=7.1.0=pypi_0
prompt-toolkit=3.0.50=pypi_0
propcache=0.3.0=pypi_0
protobuf=5.29.4=pypi_0
psutil=7.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=19.0.1=pypi_0
pycountry=24.6.1=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.2=pypi_0
python=3.11.11=he870216_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-json-logger=3.3.0=pypi_0
python-multipart=0.0.20=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.3.0=pypi_0
qwen-vl-utils=0.0.10=pypi_0
ray=2.44.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
rich-toolkit=0.13.2=pypi_0
rouge=1.0.1=pypi_0
rpds-py=0.23.1=pypi_0
ruff=0.9.10=pypi_0
s3transfer=0.11.4=pypi_0
safehttpx=0.1.6=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scikit-video=1.1.11=pypi_0
scipy=1.15.2=pypi_0
semantic-version=2.10.0=pypi_0
sentence-transformers=3.4.1=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.24.0=pypi_0
setproctitle=1.3.5=pypi_0
setuptools=75.8.0=py311h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
sse-starlette=2.2.1=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.46.1=pypi_0
swissarmytransformer=0.4.12=pypi_0
sympy=1.13.1=pypi_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.9.0=pypi_0
timm=1.0.15=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=pypi_0
tomlkit=0.13.2=pypi_0
torch=2.6.0=pypi_0
torchaudio=2.6.0=pypi_0
torchvision=0.21.0=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformer-lens=2.15.0=pypi_0
transformers=4.51.0.dev0=pypi_0
transformers-stream-generator=0.0.5=pypi_0
triton=3.2.0=pypi_0
typeguard=4.4.2=pypi_0
typer=0.15.2=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
uvloop=0.21.0=pypi_0
vllm=0.8.1=pypi_0
wadler-lindig=0.1.4=pypi_0
wandb=0.19.8=pypi_0
watchfiles=1.0.4=pypi_0
wcwidth=0.2.13=pypi_0
webdataset=0.2.111=pypi_0
websockets=15.0.1=pypi_0
wheel=0.45.1=py311h06a4308_0
xformers=0.0.29.post2=pypi_0
xgrammar=0.1.16=pypi_0
xxhash=3.5.0=pypi_0
xz=5.6.4=h5eee18b_1
yarl=1.18.3=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
